// aidbox-features/orgbac-practitioner-application/fhir.r4.sdk/types/hl7-fhir-r4-core/Patient.ts
// WARNING: This file is autogenerated by FHIR Schema Codegen.
// https://github.com/fhir-schema/fhir-schema-codegen
// Any manual changes made to this file may be overwritten.

import { Address } from './Address';
import { Attachment } from './Attachment';
import { BackboneElement } from './BackboneElement';
import { CodeableConcept } from './CodeableConcept';
import { ContactPoint } from './ContactPoint';
import { DomainResource } from './DomainResource';
import { HumanName } from './HumanName';
import { Identifier } from './Identifier';
import { Period } from './Period';
import { Reference } from './Reference';


export interface PatientCommunication extends BackboneElement {
    language: CodeableConcept;
    preferred?: boolean;
}

export interface PatientContact extends BackboneElement {
    address?: Address;
    gender?: 'male' | 'female' | 'other' | 'unknown';
    name?: HumanName;
    organization?: Reference<'Organization'>;
    period?: Period;
    relationship?: CodeableConcept[];
    telecom?: ContactPoint[];
}

export interface PatientLink extends BackboneElement {
    other: Reference<'Patient' | 'RelatedPerson'>;
    type: 'replaced-by' | 'replaces' | 'refer' | 'seealso';
}

export interface Patient extends DomainResource {
    active?: boolean;
    _active?: Element;
    address?: Address[];
    birthDate?: string;
    _birthDate?: Element;
    communication?: PatientCommunication[];
    contact?: PatientContact[];
    deceasedBoolean?: boolean;
    _deceasedBoolean?: Element;
    deceasedDateTime?: string;
    _deceasedDateTime?: Element;
    gender?: 'male' | 'female' | 'other' | 'unknown';
    _gender?: Element;
    generalPractitioner?: Reference<'Organization' | 'Practitioner' | 'PractitionerRole'>[];
    identifier?: Identifier[];
    link?: PatientLink[];
    managingOrganization?: Reference<'Organization'>;
    maritalStatus?: CodeableConcept;
    multipleBirthBoolean?: boolean;
    _multipleBirthBoolean?: Element;
    multipleBirthInteger?: number;
    _multipleBirthInteger?: Element;
    name?: HumanName[];
    photo?: Attachment[];
    telecom?: ContactPoint[];
}